requirements.txt
hydra-core==1.3.2
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.24.2
tensorflow==2.8.0
tensorflow-addons==0.20.0
tqdm